backend/requirements.txt
fastapi
uvicorn
pydantic
langchain_groq
python-dotenv
logging
